Moved console commands to subdirectories.
