+++ src/Console/Command/Module/ModuleValidateCommand.php
@@ -1,8 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace LotGD\Core\Console\Command;
+namespace LotGD\Core\Console\Command\Module;
 
+use LotGD\Core\Console\Command\BaseCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
